+++ RELEASE-NOTES.md
@@ -163,6 +163,7 @@ With the new execute API the ColorScheme class will start to play a more central
 - [#672] Need way to send errors back from subcommand. Thanks to [Garret Wilson](https://github.com/garretwilson) for raising this.
 - [#678] Exit Status section in usage help message.
 - [#680] Add annotation API for exitCodeList and exitCodeListHeading.
+- [#683] Ensure exitCodeList implementation is consistent with other usage message attributes.
 - [#575] Use mixinStandardHelpOptions in `AutoComplete$App` (add the `--version` option)
 - [#676] Bugfix: non-defined variables in `defaultValue` now correctly resolve to `null`, and options and positional parameters are now correctly considered `required` only if their default value is `null` after variable interpolation. Thanks to [ifedorenko](https://github.com/ifedorenko) for raising this.
 - [#682] Bug: incorrect evaluation for multiple occurrences of a variable.